+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # TelegramBotAPI changelog
 
+## 1.1.2
+
 ## 1.1.1
 
 * `Versions`: